requirements.txt
boto3==1.18.39
PyYAML==5.4.1